Show the largest files for each file extension
The file-by-file diff of $DIRTOSCAN is very slow. If we can move the
largest files out of it sooner then we can make the scan much faster.

If
we can move all of the files for a single extension out then we can stop
looking at that extension and, thus, also stop scanning $REFDIR.